STEM_Academy/requirements.txt
openai == 0.28.0
flask
PyPDF2
Flask 
sentence-transformers 
faiss-cpu 
numpy
flask-jwt-extended 
werkzeug